launch ostentus as part of main.py
If a main.py file doesn't exist at boot time, it will be created with
the barebones commands to import ostentus, run the init, and begin
listening on i2c.

The main.py that is created in this step is persistent and changes can
be safely made to it using Thonny.

Signed-off-by: Mike Szczys <mike@golioth.io>