separate partial update functions
Partial update can change multiple parts of the screen. This commit
breaks out the process of writing to display memory so that it is
separate form the process of executing the update.

1. Write new data to one or multiple parts of the screen using
   `partial_update_data()`
2. Execute the partial update using `partial_update_execute()`

If only one part of the screen needs to be updated, you may still use
`partial_update()` to run both parts of the process.

When more than one part of the screen is being updated, all of those
parts need to be written to memory before each update is executed, even
if some of the parts have not changed.

Signed-off-by: Mike Szczys <mike@golioth.io>